fix/login_error: resolve network disconnected on wrong credentials
Only the unauthorsied client created at the start would need to handle the network connection. The registered client creation should not
be concerned about the network connection lost while login/register. Network failure will be passed to the unregistered client observer
which is sufficient for handling the error on UI.