Fix usage of Node FFI libs to call `nfs_get_file`
C pointers have their own merit to cause confusion. On top of that,
NodeJS's `ref` package will add its own confusion.

If you are only traversing pointers, then the rules are pretty similar
to C. `obj.ref()` will "get the address of the pointer" (actually it'll
allocate a new pointer whose value is the address of the old pointer)
and `obj.deref()` will dereference the pointer (i.e. get the pointee).

So, if `foo` was created calling `ref.alloc(ref.refType('int'))`, you
can treat as if it were a C `int*`. However, you always manipulate
references and `foo.deref()` will hold not an `int`, but a `Buffer`
whose size is the size of an `int` and whose contents represent the
value contained by this `int`. To convert the `Buffer` into a JS Number
value, you need to call `deref` once more (i.e. `foo.deref().deref()`).

The big confusion added by the `ref` package is the fact that you always
pass the address of the buffer to the C function. If `foo` is a
"pointer", the address of the pointer (represented by NodeJS as a
`Buffer` size_t-sized), and not the memory address being hold by the
pointer, is passed when you call a C function.

There is also the confusion of `ref-struct` package.
`ref.refType(FileDetails)` will create a type without the helper
functions present in `FileDetails`. So you need to create a new
`FileDetails` instance passing the backing `Buffer` as argument once you
obtain the `FileDetails` navigating through references.