Fix FFI usage of appendable_data_filter_type
C/FFI interface for C's `enum` is `int`, not a JS object. Update code to
receive such `int` and convert to a proper `Enum` within JS environment.

We use the `String` extracted from `EnumItem` instead the `EnumItem`
itself because the rest of the code is expecting a string which will be
sent over the wire.